Improvised Bash Installation Script...

+++ src/aur.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 
 paru=(
 	hyprshot
@@ -11,33 +12,36 @@ paru=(
 
 while true;do
 
-echo 
-echo 
+if [ -f /usr/bin/paru ]; then 
+	break 
+else 
+	echo 
+	echo 
+	read -rp " :: Do u wanna install the paru aur helper? (y/n): " aur
 
-read -p " :: Do u wanna install the paru aur helper? (y/n): " aur
+	case "$aur" in
+		[y*])
+			echo " :: Installing paru ( AUR helper ) ::"
+			git clone https://aur.archlinux.org/paru.git
+			cd paru/ || return
+			makepkg -si 
+			sudo rm -r ~/paru/
+			cd ~ || return
+			break
+			;;
 
-case "$aur" in
-	[y*])
-		echo " :: Installing paru ( AUR helper ) ::"
-		git clone https://aur.archlinux.org/paru.git
-		cd paru/
-		makepkg -si 
-		sudo rm -r ~/paru/
-		cd ~
-		break
-		;;
+		[n]*)
+			break
+			;;
 
-	[n]*)
-		break
-		;;
+		*)
+			echo 
+			echo
+			echo " :: Please answer yes(y) or no(n) ::"
+			;;
 
-	*)
-		echo 
-		echo
-		echo " :: Please answer yes(y) or no(n) ::"
-		;;
-
-esac
+		esac
+	fi
 
 done
 
@@ -47,7 +51,7 @@ echo
 
 echo
 
-read -p " :: Do u wanna install some aur packages? (y/n): " depen
+read -rp " :: Do u wanna install some aur packages? (y/n): " depen
 
 case "$depen" in
 	[y]*)
